corrected the newsletter functions
corrected the newsletter function to remove the subscribe requirement to have a token. I am now the only user who can send out the newsletter as my userid is the one that is checked in the if statement.  Added a statement about load times on the petform. I swear I am done Serofina can